Tasks tables (queued and history) have been created. Header for businesses shows page for each business.
Tasks tables are now divided from pending and completed. After a business is deleted, the tasks for that business are also deleted. Able to click on the business shown in the header and it displays the dedicated page. Each dedicated page shows the queued task list and graph (for # of tasks per month.